fix(metrics): stop printing a 32-bit SIZE_T through %llu as 8.7 exabytes of VRAM
DXGI declares DedicatedVideoMemory as SIZE_T, which is four bytes in the
32-bit hook. Passing it straight to a %llu conversion read four bytes of
the neighbouring stack as the high half, which is where Gothic II session
20260916_013230's 'VRAM Total: 8760129288574861312 bytes' came from; the
value CE stored and reported was always the correct 11943 MB.

The adapter descriptor the same loop matched its LUID against was also
uninitialised with GetDesc1's result unchecked, so a failed call left
every field - including that LUID - in play.

+++ hook/wrappers/wrapper_hooks_devices.cpp
@@ -74,7 +74,8 @@ HRESULT WINAPI Wrapped_D3D12CreateDevice(IUnknown* pAdapter, D3D_FEATURE_LEVEL M
                 WrapperLog(
                     "Wrapper: D3D12CreateDevice - Adapter LUID: %08X:%08X, "
                     "VRAM: %llu MB",
-                    desc.AdapterLuid.HighPart, desc.AdapterLuid.LowPart, desc.DedicatedVideoMemory / (1024 * 1024));
+                    desc.AdapterLuid.HighPart, desc.AdapterLuid.LowPart,
+                    static_cast<unsigned long long>(desc.DedicatedVideoMemory / (1024 * 1024)));
             }
             pDXGIAdapter->Release();
         }
